+++ src/components/details/details.styles.ts
@@ -34,7 +34,8 @@ export default css`
   }
 
   .details__header${focusVisibleSelector} {
-    box-shadow: var(--sl-focus-ring);
+    outline: var(--sl-focus-ring);
+    outline-offset: calc(1px + var(--sl-focus-ring-offset));
   }
 
   .details--disabled .details__header {