+++ src/components/image-comparer/image-comparer.styles.ts
@@ -71,7 +71,7 @@ export default css`
   }
 
   .image-comparer__handle${focusVisibleSelector} {
-    outline: none;
-    box-shadow: var(--sl-focus-ring);
+    outline: var(--sl-focus-ring);
+    outline-offset: var(--sl-focus-ring-offset);
   }
 `;